--- packages/web/app/[team]/[project]/new-table/layout.tsx
@@ -1,17 +1,0 @@
-import NavNewTable from "@/components/nav-new-table";
-import SubNavHeader from "@/components/sub-nav-header";
-
-export default async function NewTableLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <div>
-      <SubNavHeader>
-        <NavNewTable crumbTitle="New Table" />
-      </SubNavHeader>
-      {children}
-    </div>
-  );
-}
